MAINT: rename odlpp to odlcuda
