+++ odlcuda/cuda/CudaVectorImpl.cu
@@ -19,9 +19,9 @@
 #include <iostream>
 
 // ODL
-#include <odlpp/cuda/DeviceVectorImpl.h>
-#include <odlpp/cuda/CudaVectorImpl.h>
-#include <odlpp/cuda/TypeMacro.h>
+#include <odlcuda/cuda/DeviceVectorImpl.h>
+#include <odlcuda/cuda/CudaVectorImpl.h>
+#include <odlcuda/cuda/TypeMacro.h>
 
 // Utils
 #include <odl_cpp_utils/cuda/thrustUtils.h>
