+++ odlcuda/cuda/SliceHelper.h
