+++ odlcuda/cuda/dummy.cpp
